Expose fs API as part of module API
To make this module easier to use in a forward-compatible way, include
the fs module properties through this module via the prototype chain.
This allows fs-file-sync-fd to be used as the fs module in code which
requires it.  This makes replacing fs-file-sync-fd with fs when moving
to Node v6 or later very easy and painless.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>